Set Arrow Down Icon to Black
Given that the BoxDecoration for the opn button is explicitly set to Colors.white, this means the arrow cannot be seen when using
flutter's Material dark mode as the icon also appears white. Therefore explicitly setting the icon to black ensures it can always be seen regardless of theme settings.